Don't include testing helpers in release code
