+++ runtime/calamari/src/xcm_config.rs
@@ -335,42 +335,3 @@ impl orml_xtokens::Config for Runtime {
     type MultiLocationsFilter = AssetManager;
     type ReserveProvider = AbsoluteReserveProvider;
 }
-
-#[cfg(test)]
-mod production_xcm_weigher_tests {
-    use super::{Call, XcmExecutorConfig};
-    use manta_primitives::xcm::{
-        self_reserve_xcm_message_receiver_side, self_reserve_xcm_message_sender_side,
-        to_reserve_xcm_message_receiver_side, to_reserve_xcm_message_sender_side,
-        ADVERTISED_DEST_WEIGHT,
-    };
-    use xcm_executor::traits::WeightBounds;
-
-    #[test]
-    fn test_receiver_side_weights() {
-        let weight = <XcmExecutorConfig as xcm_executor::Config>::Weigher::weight(
-            &mut self_reserve_xcm_message_receiver_side::<Call>(),
-        )
-        .unwrap();
-        assert!(weight <= ADVERTISED_DEST_WEIGHT);
-
-        let weight = <XcmExecutorConfig as xcm_executor::Config>::Weigher::weight(
-            &mut to_reserve_xcm_message_receiver_side::<Call>(),
-        )
-        .unwrap();
-        assert!(weight <= ADVERTISED_DEST_WEIGHT);
-    }
-
-    #[test]
-    fn test_sender_side_xcm_weights() {
-        let mut msg = self_reserve_xcm_message_sender_side::<Call>();
-        let weight =
-            <XcmExecutorConfig as xcm_executor::Config>::Weigher::weight(&mut msg).unwrap();
-        assert!(weight < ADVERTISED_DEST_WEIGHT);
-
-        let mut msg = to_reserve_xcm_message_sender_side::<Call>();
-        let weight =
-            <XcmExecutorConfig as xcm_executor::Config>::Weigher::weight(&mut msg).unwrap();
-        assert!(weight < ADVERTISED_DEST_WEIGHT);
-    }
-}
