+++ runtime/calamari/tests/xcm_mock/relay_chain.rs
@@ -16,6 +16,8 @@
 
 //! Relay chain runtime mock.
 
+#![cfg(test)]
+
 use frame_support::{
     construct_runtime, parameter_types,
     traits::{ConstU32, Everything, Nothing},
